# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user profile",
                "description": "Returns the authenticated user's profile information including username, isPublic status, and meta data.",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Current user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserShowPrivate"
                                },
                                "examples": {
                                    "user_profile": {
                                        "summary": "Current user profile",
                                        "value": {
                                            "username": "janedoe",
                                            "isPublic": true,
                                            "meta": {
                                                "theme": "dark",
                                                "language": "en"
                                            },
                                            "account": {
                                                "@iri": "https://bookmarkhive.test/profile/janedoe",
                                                "username": "janedoe",
                                                "instance": "bookmarkhive.test",
                                                "inboxUrl": "https://bookmarkhive.test/profile/janedoe/inbox",
                                                "outboxUrl": "https://bookmarkhive.test/profile/janedoe/outbox",
                                                "sharedInboxUrl": "https://bookmarkhive.test/inbox",
                                                "followerUrl": "https://bookmarkhive.test/profile/janedoe/followers",
                                                "followingUrl": "https://bookmarkhive.test/profile/janedoe/following"
                                            },
                                            "@iri": "https://bookmarkhive.test/users/me"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete current user account",
                "description": "Permanently deletes the authenticated user's account and all associated data. After deletion, accessing GET /users/me will return 404.",
                "operationId": "deleteCurrentUser",
                "responses": {
                    "204": {
                        "description": "User account deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update current user profile",
                "description": "Updates the authenticated user's profile. Username and password changes will invalidate existing JWT tokens. Meta data is merged, not replaced.",
                "operationId": "updateCurrentUser",
                "requestBody": {
                    "description": "User update data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "New username",
                                        "type": "string",
                                        "maxLength": 32,
                                        "minLength": 3
                                    },
                                    "password": {
                                        "description": "New password",
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "isPublic": {
                                        "description": "Whether the profile is public",
                                        "type": "boolean"
                                    },
                                    "meta": {
                                        "description": "Additional metadata as key-value pairs (merged with existing)",
                                        "type": "object",
                                        "additionalProperties": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "422": {
                        "description": "Validation error - invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_data": {
                                        "summary": "Validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "description": "Error object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "description": "HTTP status code",
                                "type": "integer"
                            },
                            "message": {
                                "description": "Error message",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "AccountShowPublic": {
                "description": "Account object representing an ActivityPub account",
                "properties": {
                    "@iri": {
                        "description": "IRI of the account resource",
                        "type": "string",
                        "format": "iri"
                    },
                    "username": {
                        "description": "Account username",
                        "type": "string"
                    },
                    "instance": {
                        "description": "Instance domain where the account is hosted",
                        "type": "string"
                    },
                    "inboxUrl": {
                        "description": "ActivityPub inbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "outboxUrl": {
                        "description": "ActivityPub outbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "sharedInboxUrl": {
                        "description": "ActivityPub shared inbox URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "followerUrl": {
                        "description": "ActivityPub followers collection URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "followingUrl": {
                        "description": "ActivityPub following collection URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserShowPrivate": {
                "description": "User object with owner-level details",
                "properties": {
                    "username": {
                        "description": "Username",
                        "type": "string"
                    },
                    "isPublic": {
                        "description": "Whether the profile is public",
                        "type": "boolean"
                    },
                    "meta": {
                        "description": "Additional metadata",
                        "type": "object",
                        "additionalProperties": true
                    },
                    "account": {
                        "$ref": "#/components/schemas/AccountShowPublic"
                    },
                    "@iri": {
                        "description": "IRI of the user resource",
                        "type": "string",
                        "format": "iri"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}